pyproject.toml
[tool.poetry]
name = "alone_memo"
version = "0.1.0"
description = ""
authors = ["nbgh46 <nbgh46@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
pymongo = "^3.11.4"
requests = "^2.25.1"
bs4 = "^0.0.1"
PyJWT = "^2.1.0"
python-dotenv = "^0.17.1"
pytest-flask = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
